requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.5.1
South==1.0
django-appconf==0.6
django-compressor==1.4
six==1.7.3
wsgiref==0.1.2
httplib2==0.9
python-instagram==1.1.3
simplejson==3.6.3